requirements.txt
openvino>=2024.3.0
optimum-intel>=1.17.0
transformers>=4.44.0
accelerate>=0.33.0
datasets>=2.19.0
nncf>=2.10.0
rich>=13.7.0
torch>=2.3.0
